+++ web/src/features/annotation-queues/components/CreateOrEditAnnotationQueueButton.tsx
@@ -117,6 +117,7 @@ export const CreateOrEditAnnotationQueueButton = ({
   const configsData = api.scoreConfigs.all.useQuery(
     {
       projectId,
+      filter: [],
     },
     {
       enabled: hasQueueAccess && isOpen,
