+++ web/src/features/score-configs/components/ScoreConfigDetails.tsx
