+++ web/src/features/score-configs/components/ScoreConfigSettings.tsx
